refactor: 타입 변경 #41
interest 필드 double에서 BigDecimal로 변경

+++ src/main/java/cheongsan/domain/simulator/dto/LoanDTO.java
@@ -17,7 +17,7 @@ public class LoanDTO {
     private Long id;
     private String loanName; // 대출명
     private BigDecimal principal; // 남은 원금
-    private double interestRate; // 연이율(소수점)
+    private BigDecimal interestRate; // 연이율(소수점)
     private BigDecimal monthlyPayment; // 월 납입금
     private LocalDate startDate; // 대출 시작일
     private LocalDate endDate; // 대출 만기일
